Corrected some phrasing
A couple of comments:
Line 313: "It could be used for a base instrument" - should this be "bass instrument"? I wasn't sure about this so I didn't change it, but if it is not supposed to be "bass" then I don't understand what it means.
Line 325: Technically there is a difference between arpeggios and scales, so I changed "scales" to "notes"
Line 494: Since those patterns are not actually being combined (i.e. joined or concatenated) I changed it to "reduced"